+++ src/main/java/dev/wgrgwg/somniverse/member/service/AuthService.java
@@ -3,11 +3,16 @@ package dev.wgrgwg.somniverse.member.service;
 import dev.wgrgwg.somniverse.global.exception.CustomException;
 import dev.wgrgwg.somniverse.member.domain.Member;
 import dev.wgrgwg.somniverse.member.domain.RefreshToken;
+import dev.wgrgwg.somniverse.member.dto.request.LoginRequest;
 import dev.wgrgwg.somniverse.member.dto.response.TokenResponse;
 import dev.wgrgwg.somniverse.member.exception.MemberErrorCode;
 import dev.wgrgwg.somniverse.member.repository.RefreshTokenRepository;
 import dev.wgrgwg.somniverse.security.jwt.provider.JwtProvider;
+import dev.wgrgwg.somniverse.security.userdetails.CustomUserDetails;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -15,8 +20,29 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class AuthService {
 
+    private final AuthenticationManager authenticationManager;
     private final JwtProvider jwtProvider;
     private final RefreshTokenRepository refreshTokenRepository;
+
+    @Transactional
+    public TokenResponse login(LoginRequest loginRequest) {
+        UsernamePasswordAuthenticationToken authToken = new UsernamePasswordAuthenticationToken(
+            loginRequest.email(), loginRequest.password());
+        Authentication authentication = authenticationManager.authenticate(authToken);
+
+        CustomUserDetails userDetails = (CustomUserDetails) authentication.getPrincipal();
+        Member member = userDetails.getMember();
+
+        TokenResponse tokenResponse = jwtProvider.generateToken(member);
+
+        RefreshToken refreshToken = RefreshToken.builder()
+            .member(member)
+            .value(tokenResponse.refreshToken())
+            .build();
+        refreshTokenRepository.save(refreshToken);
+
+        return tokenResponse;
+    }
 
     @Transactional
     public TokenResponse reissue(String refreshToken) {
@@ -37,5 +63,10 @@ public class AuthService {
         refreshTokenFromDB.updateValue(newTokens.refreshToken());
 
         return new TokenResponse(newTokens.accessToken(), newTokens.refreshToken());
+    }
+
+    @Transactional
+    public void logout(String refreshToken) {
+        refreshTokenRepository.findByValue(refreshToken).ifPresent(refreshTokenRepository::delete);
     }
 }
